backend/requirements.txt
amqp==5.3.1
asgiref==3.11.1
attrs==26.1.0
autobahn==25.12.2
Automat==25.4.16
billiard==4.2.4
cbor2==5.8.0
celery==5.6.2
certifi==2026.2.25
cffi==2.0.0
channels==4.3.2
charset-normalizer==3.4.6
click==8.3.1
click-didyoumean==0.3.1
click-plugins==1.1.1.2
click-repl==0.3.0
constantly==23.10.4
cryptography==46.0.5
daphne==4.2.1
Django==6.0.3
django-cors-headers==4.9.0
django_fernet_encrypted_fields==0.3.1
djangorestframework==3.16.1
djangorestframework-simplejwt==5.3.1
docker==7.1.0
hyperlink==21.0.0
idna==3.11
Incremental==24.11.0
iniconfig==2.3.0
kombu==5.6.2
msgpack==1.1.2
packaging==26.0
pluggy==1.6.0
prompt_toolkit==3.0.52
psycopg==3.2.6
psycopg-binary==3.2.6
psycopg2-binary==2.9.11
py-ubjson==0.16.1
pyasn1==0.6.3
pyasn1_modules==0.4.2
pycparser==3.0
Pygments==2.19.2
PyJWT==2.12.1
pyOpenSSL==26.0.0
pytest==9.0.2
pytest-asyncio==1.3.0
pytest-django==4.12.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
requests==2.32.5
service-identity==24.2.0
six==1.17.0
sqlparse==0.5.5
Twisted==25.5.0
txaio==25.12.2
typing_extensions==4.15.0
tzdata==2025.3
tzlocal==5.3.1
ujson==5.12.0
urllib3==2.6.3
vine==5.1.0
wcwidth==0.6.0
zope.interface==8.2
